requirements.txt
python-telegram-bot[job-queue]>=20.7
requests==2.31.0
flask==3.0.0
pytz==2023.3
python-dotenv

